refactor: move component communication to infra crate
commit-id:4c9d9914
