+++ crates/mempool_infra/src/component_definitions.rs
@@ -1,11 +1,30 @@
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-use tokio::sync::mpsc::Sender;
+use tokio::sync::mpsc::{Receiver, Sender};
 
 #[async_trait]
 pub trait ComponentRequestHandler<Request, Response> {
     async fn handle_request(&mut self, request: Request) -> Response;
+}
+
+pub struct ComponentCommunication<T: Send + Sync> {
+    tx: Option<Sender<T>>,
+    rx: Option<Receiver<T>>,
+}
+
+impl<T: Send + Sync> ComponentCommunication<T> {
+    pub fn new(tx: Option<Sender<T>>, rx: Option<Receiver<T>>) -> Self {
+        Self { tx, rx }
+    }
+
+    pub fn take_tx(&self) -> Sender<T> {
+        self.tx.to_owned().expect("Sender should be available, could be taken only once")
+    }
+
+    pub fn take_rx(&mut self) -> Receiver<T> {
+        self.rx.take().expect("Receiver should be available, could be taken only once")
+    }
 }
 
 pub struct ComponentRequestAndResponseSender<Request, Response>
