+++ crates/mempool_node/Cargo.toml
@@ -13,15 +13,15 @@ anyhow.workspace = true
 clap.workspace = true
 const_format.workspace = true
 futures.workspace = true
-starknet_gateway = { path = "../gateway", version = "0.0" }
-starknet_mempool_infra = { path = "../mempool_infra", version = "0.0" }
-starknet_mempool = { path = "../mempool", version = "0.0" }
-starknet_mempool_types = { path = "../mempool_types", version = "0.0" }
-serde.workspace = true
 papyrus_config.workspace = true
+serde.workspace = true
+starknet_gateway = { path = "../gateway", version = "0.0" }
+starknet_mempool = { path = "../mempool", version = "0.0" }
+starknet_mempool_infra = { path = "../mempool_infra", version = "0.0" }
+starknet_mempool_types = { path = "../mempool_types", version = "0.0" }
 tokio.workspace = true
-tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
+tracing.workspace = true
 validator.workspace = true
 
 [dev-dependencies]
@@ -30,4 +30,4 @@ assert_matches.workspace = true
 colored.workspace = true
 pretty_assertions.workspace = true
 serde_json.workspace = true
-test_utils = {path = "../test_utils"}
+test_utils = { path = "../test_utils" }
