+++ crates/mempool_node/src/communication.rs
@@ -1,25 +1,12 @@
 use std::sync::Arc;
 
+use starknet_mempool_infra::component_definitions::ComponentCommunication;
 use starknet_mempool_types::communication::{
     MempoolClientImpl, MempoolRequestAndResponseSender, SharedMempoolClient,
 };
 use tokio::sync::mpsc::{channel, Receiver, Sender};
 
 use crate::config::MempoolNodeConfig;
-
-pub struct ComponentCommunication<T: Send + Sync> {
-    tx: Option<Sender<T>>,
-    rx: Option<Receiver<T>>,
-}
-
-impl<T: Send + Sync> ComponentCommunication<T> {
-    fn take_tx(&self) -> Sender<T> {
-        self.tx.to_owned().expect("Sender should be available, could be taken only once")
-    }
-    fn take_rx(&mut self) -> Receiver<T> {
-        self.rx.take().expect("Receiver should be available, could be taken only once")
-    }
-}
 
 pub struct MempoolNodeCommunication {
     mempool_channel: ComponentCommunication<MempoolRequestAndResponseSender>,
@@ -39,7 +26,7 @@ pub fn create_node_channels() -> MempoolNodeCommunication {
     let (tx_mempool, rx_mempool) =
         channel::<MempoolRequestAndResponseSender>(MEMPOOL_INVOCATIONS_QUEUE_SIZE);
     MempoolNodeCommunication {
-        mempool_channel: ComponentCommunication { tx: Some(tx_mempool), rx: Some(rx_mempool) },
+        mempool_channel: ComponentCommunication::new(Some(tx_mempool), Some(rx_mempool)),
     }
 }
